Add missing license header to Rust files
And update Makefile to ensure they are checked

+++ lib/srv/desktop/rdp/rdpclient/src/lib.rs
@@ -1,3 +1,17 @@
+// Copyright 2021 Gravitational, Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 pub mod errors;
 pub mod piv;
 pub mod rdpdr;
@@ -326,7 +340,7 @@ fn wait_for_fd(fd: usize) -> bool {
 /// client_ptr must be a valid pointer to a Client.
 /// handle_bitmap *must not* free the memory of CGOBitmap.
 #[no_mangle]
-pub unsafe extern "C" fn read_rdp_output(client_ptr: *mut Client, client_ref: i64) -> CGOError {
+pub unsafe extern "C" fn read_rdp_output(client_ptr: *mut Client, client_ref: usize) -> CGOError {
     let client = Client::from_ptr(client_ptr);
     let client = match client {
         Some(client) => client,
